fix increments & specify laravel version

+++ database/migrations/2019_05_25_141846_create_products_table.php
@@ -14,9 +14,9 @@ class CreateProductsTable extends Migration
     public function up()
     {
         Schema::create('products', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 			$table->string('name');
-			$table->integer('amount');
+			$table->bigInteger('amount');
             $table->timestamps();
         });
     }
